[POC] Class-based scaffolding for structured functions
Previously, implementations of structured functions were still defined
as regular functions (one function for meta, one function for
the out implementation).  This POC changes it so that the meta
function is the constructor for a class, and the impl function is
a method on the class.  This means it is now possible to implicitly
pass more information from meta function to the real function;
the biggest use case for this is TensorIterator.  There isn't enough
scaffolding yet to actually use this for that case, but it's enough
to flip the current code over.

There is one slight performance downside to the current formulation,
which is we have to default initialize TensorMeta before it gets
assigned by the constructor.  Not sure if compiler can optimize
this away.  However, we could get rid of this problem by virtualizing
allocation/resizing into a method of some sort; left to future
work.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: b08f74d854c8117b846cc322752114327ef5ac28
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48262